fix #4846 address tournament review comments
